backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlmodel
sentence-transformers
scikit-learn
python-dotenv
pdfplumber
unstructured
python-docx

